requirements.txt
lxml
numpy
pillow
requests
easyocr
bs4
